+++ codex-rs/tui/src/lib.rs
@@ -49,6 +49,7 @@ mod shimmer;
 mod slash_command;
 mod status_indicator_widget;
 mod streaming;
+mod string_utils;
 mod text_formatting;
 mod tui;
 mod user_approval_widget;
